py/requirements.txt
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
cffi==1.17.1
memory-profiler==0.61.0
psutil==6.1.0
pycparser==2.22
